refactor: extract base.py from networks/_legacy_full (2/N)
