+++ stisim/networks/__init__.py
@@ -1,4 +1,3 @@
 """STIsim networks package — split across base, mf, fsw, msm, layered_networks, matchers."""
-
-# Temporary during refactor; all symbols still come from _legacy_full
-from ._legacy_full import *  # noqa: F401, F403
+from .base import NoPartnersFound, BasePars, NetworkPars, BaseNetwork
+from ._legacy_full import *  # noqa: F401, F403 (remaining classes during refactor)
